+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## Ver 1.0.2  
+* Fixed an issue where Clive stopped flying after doing to many actions in the air
+* Fixed an issue where Megaflare dodge only worked once in the air
+* You can now hold L3 or Alt to enter "Flight Mode", where attack/magic will allow you to fly up or down
+
+
 ## Ver 1.0.1  
 * Allows better movment control while flying